Changed interface of allocation function to rely on size_t rather than unsigned integer; include stddef.h at frontend-level (size_t). Raised threshold until using the largest alignment (LIBXSMM_ALLOC_ALIGNFCT). Attempted to support huge pages under Windows (libxsmm_allocate); appears to always fail when running with regular permissions. Code cleanup and more comments in allocation layer. Enabled low-level memory allocation for all allocations (this is just to run it one time through all regression tests).

+++ src/libxsmm_alloc.h
@@ -44,24 +44,24 @@ typedef enum libxsmm_alloc_flags {
   LIBXSMM_ALLOC_FLAG_DEFAULT = LIBXSMM_ALLOC_FLAG_RW
 } libxsmm_alloc_flags;
 
-LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE unsigned int libxsmm_gcd(unsigned int a, unsigned int b);
-LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE unsigned int libxsmm_lcm(unsigned int a, unsigned int b);
-LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE unsigned int libxsmm_alignment(unsigned int size, unsigned int alignment);
+LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE size_t libxsmm_gcd(size_t a, size_t b);
+LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE size_t libxsmm_lcm(size_t a, size_t b);
+LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE size_t libxsmm_alignment(size_t size, size_t alignment);
 
 /** Receive the size, the flags, or the extra attachment of the given buffer. */
-LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_alloc_info(const void* memory, unsigned int* size, int* flags, void** extra);
+LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_alloc_info(const void* memory, size_t* size, int* flags, void** extra);
 
 /** Allocate memory of the requested size, which is aligned according to the given alignment. */
-LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_allocate(void** memory, unsigned int size, unsigned int alignment, int flags,
+LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_allocate(void** memory, size_t size, size_t alignment, int flags,
   /* The extra information is stored along with the allocated chunk; can be NULL/zero. */
-  const void* extra, unsigned int extra_size);
+  const void* extra, size_t extra_size);
 LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_deallocate(const void* memory);
 
 /** Attribute memory allocation such as to revoke protection flags. */
 LIBXSMM_EXTERN_C LIBXSMM_RETARGETABLE int libxsmm_alloc_attribute(const void* memory, int flags, const char* name);
 
 /** Allocate memory (malloc/free interface). */
-LIBXSMM_INLINE_EXPORT LIBXSMM_RETARGETABLE void* libxsmm_malloc(unsigned int size)
+LIBXSMM_INLINE_EXPORT LIBXSMM_RETARGETABLE void* libxsmm_malloc(size_t size)
 #if defined(LIBXSMM_BUILD)
 ;
 #else
